feat: extract auth logic into new beam-auth crate
Create beam-auth package with two feature flags:
- utils: AuthService trait + LocalAuthService, SessionStore trait +
  RedisSessionStore, UserRepository trait + SqlUserRepository, and
  all supporting domain types (User, CreateUser, AuthenticatedUser,
  AuthResponse, Claims, etc.)
- server: Salvo HTTP handlers for register/login/refresh/logout,
  using Arc<dyn AuthService> injected via depot

Remove the corresponding modules from beam-stream (services/auth,
services/session_store, repositories/user, models/domain/user,
routes/auth) and update imports throughout to use beam-auth.

Auth routes in beam-stream now call beam_auth::server::auth_routes()
with Arc<dyn AuthService> injected separately into the depot so the
handlers remain fully decoupled from AppState.
