+++ beam-auth/src/lib.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "utils")]
+pub mod utils;
+
+#[cfg(feature = "server")]
+pub mod server;
