+++ beam-auth/src/server/mod.rs
@@ -1,0 +1,3 @@
+pub mod routes;
+
+pub use routes::auth_routes;
