+++ beam-auth/src/utils/models.rs
