+++ beam-stream/src/models/domain/mod.rs
@@ -3,7 +3,6 @@ pub mod library;
 pub mod movie;
 pub mod show;
 pub mod stream;
-pub mod user;
 
 // Re-export all domain model types for convenience
 pub use file::*;
@@ -11,4 +10,3 @@ pub use library::*;
 pub use movie::*;
 pub use show::*;
 pub use stream::*;
-pub use user::*;
