+++ beam-stream/src/services/mod.rs
@@ -1,9 +1,7 @@
-pub mod auth;
 pub mod hash;
 pub mod library;
 pub mod media_info;
 pub mod metadata;
-pub mod session_store;
 pub mod transcode;
 
 // Re-export types for convenience
