+++ beam-stream/src/state.rs
@@ -2,14 +2,18 @@ use sea_orm::DatabaseConnection;
 use std::ops::Deref;
 use std::sync::Arc;
 
+use beam_auth::utils::{
+    repository::SqlUserRepository,
+    service::{AuthService, LocalAuthService},
+    session_store::{RedisSessionStore, SessionStore},
+};
+
 use crate::{
     config::ServerConfig,
     services::{
-        auth::{AuthService, LocalAuthService},
         hash::{HashConfig, HashService, LocalHashService},
         library::{LibraryService, LocalLibraryService},
         metadata::{MetadataConfig, MetadataService, StubMetadataService},
-        session_store::{RedisSessionStore, SessionStore},
         transcode::{LocalTranscodeService, TranscodeService},
     },
 };
@@ -87,7 +91,7 @@ impl AppServices {
         let stream_repo = Arc::new(crate::repositories::SqlMediaStreamRepository::new(
             db.clone(),
         ));
-        let user_repo = Arc::new(crate::repositories::SqlUserRepository::new(db.clone()));
+        let user_repo = Arc::new(SqlUserRepository::new(db.clone()));
 
         let hash_service = Arc::new(LocalHashService::new(hash_config));
         let media_info_service =
@@ -107,7 +111,7 @@ impl AppServices {
         let auth_service = Arc::new(LocalAuthService::new(
             user_repo,
             session_store.clone(),
-            config.clone(),
+            config.jwt_secret.clone(),
         ));
 
         Self {
